fix: исправлена логика слоев приложения

+++ src/main/java/ru/practicum/shareit/item/service/ItemServiceImpl.java
@@ -1,27 +1,25 @@
 package ru.practicum.shareit.item.service;
 
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.practicum.shareit.error.ItemNotExistsExeption;
 import ru.practicum.shareit.error.PermissionException;
+import ru.practicum.shareit.error.UserDoesNotExixtsException;
 import ru.practicum.shareit.item.dto.ItemDto;
 import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.item.storage.ItemRepository;
+import ru.practicum.shareit.user.service.UserService;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@AllArgsConstructor
 public class ItemServiceImpl implements ItemService {
     private final ItemRepository itemStorage;
-    private long id;
-
-    public ItemServiceImpl(ItemRepository itemStorage) {
-        this.itemStorage = itemStorage;
-        this.id = 1L;
-    }
+    private final UserService userService;
 
     @Override
     public Item getItem(Long id) {
@@ -31,6 +29,9 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public List<Item> getItems(Long ownerId) {
+        if (userService.getUser(ownerId) == null) {
+            throw new UserDoesNotExixtsException(ownerId.toString());
+        }
 
         return itemStorage.getAll().stream()
                 .filter(item -> Objects.equals(item.getOwnerId(), ownerId))
@@ -39,10 +40,6 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public List<Item> findItems(String text) {
-        if (text.isBlank()) {
-            return Collections.emptyList();
-        }
-
         return itemStorage.getAll().stream()
                 .filter(item ->
                         (item.getName().toLowerCase().contains(text.toLowerCase())
@@ -53,8 +50,11 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Item createItem(ItemDto itemDto, Long ownerId) {
+        if (userService.getUser(ownerId) == null) {
+            throw new UserDoesNotExixtsException(ownerId.toString());
+        }
+
         Item item = Item.builder()
-                .id(id++)
                 .name(itemDto.getName())
                 .description(itemDto.getDescription())
                 .available(true)
@@ -66,6 +66,10 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Item updateItem(ItemDto itemDto, Long itemId, Long ownerId) {
+        if (userService.getUser(ownerId) == null) {
+            throw new UserDoesNotExixtsException(ownerId.toString());
+        }
+
         Optional<Item> itemOptional = Optional.ofNullable(itemStorage.get(itemId));
         if (itemOptional.isEmpty()) {
             throw new ItemNotExistsExeption(itemId.toString());
@@ -86,6 +90,10 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public boolean deleteItem(Long id, Long ownerId) {
+        if (userService.getUser(ownerId) == null) {
+            throw new UserDoesNotExixtsException(ownerId.toString());
+        }
+
         if (Objects.equals(itemStorage.get(id).getOwnerId(), ownerId)) {
             return itemStorage.delete(id);
         } else {
